@@ -15,24 +15,24 @@
     "url": "https://github.com/cmcampione/thingshub.org.git"
   },
   "dependencies": {
-    "axios": "^0.18.0",
+    "axios": "^0.18.1",
     "g": "^2.0.1",
     "jwt-decode": "^2.2.0",
-    "npm": "^6.1.0",
+    "npm": "^6.9.0",
     "promise.prototype.finally": "^3.1.0",
-    "qs": "^6.5.2",
-    "socket.io-client": "^2.1.1"
+    "qs": "^6.7.0",
+    "socket.io-client": "^2.2.0"
   },
   "devDependencies": {
     "@types/jwt-decode": "^2.2.1",
-    "@types/promise.prototype.finally": "^2.0.2",
+    "@types/promise.prototype.finally": "^2.0.3",
     "@types/socket.io-client": "^1.4.32",
-    "dts-bundle-generator": "^1.6.0",
-    "filemanager-webpack-plugin": "^1.0.27",
-    "ts-loader": "^4.4.2",
-    "webpack": "^4.15.1",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-server": "^3.1.5",
+    "dts-bundle-generator": "^1.6.1",
+    "filemanager-webpack-plugin": "^2.0.5",
+    "ts-loader": "^4.5.0",
+    "webpack": "^4.35.0",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2",
     "webpack-shell-plugin": "^0.5.0"
   }
 }
